requirements.txt
absl-py==2.1.0
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiohttp-retry==2.9.1
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.8.0
appdirs==1.4.4
argcomplete==3.5.3
asyncssh==2.19.0
atpublic==5.0
attrs==24.3.0
beautifulsoup4==4.12.3
billiard==4.2.1
blinker==1.9.0
boto3==1.35.95
botocore==1.35.95
Brotli==1.1.0
cachetools==5.5.0
celery==5.4.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloud-ml-common==0.0.2
colorama==0.4.6
ConfigArgParse==1.7
configobj==5.0.9
contourpy==1.3.1
cryptography==43.0.3
cycler==0.12.1
dacite==1.7.0
decorator==5.1.1
Deprecated==1.2.15
dictdiffer==0.9.0
dill==0.3.9
diskcache==5.6.3
distro==1.9.0
dnspython==2.7.0
docker-pycreds==0.4.0
dpath==2.2.0
dulwich==0.22.7
dvc==3.59.0
dvc-data==3.16.7
dvc-gdrive==3.0.1
dvc-gs==3.0.1
dvc-http==2.32.0
dvc-objects==5.1.0
dvc-render==1.0.2
dvc-studio-client==0.21.0
dvc-task==0.40.2
entrypoints==0.4
fastapi==0.115.6
fiftyone==1.2.0
fiftyone-brain==0.18.2
fiftyone_db==1.1.7
filelock==3.16.1
Flask==3.1.0
Flask-Cors==5.0.0
Flask-Login==0.6.3
flatten-dict==0.4.2
flufl.lock==8.1.0
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.12.0
ftfy==6.3.1
funcy==2.0
future==1.0.0
gcsfs==2024.12.0
gevent==24.11.1
geventhttpclient==2.3.3
gitdb==4.0.12
GitPython==3.1.44
glob2==0.7
google-api-core==2.24.0
google-api-python-client==2.159.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==0.4.6
google-cloud-core==2.4.1
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grandalf==0.8
graphql-core==3.2.5
greenlet==3.1.1
grpcio==1.69.0
gto==1.7.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
humanize==4.11.0
hydra-core==1.3.0
Hypercorn==0.17.3
hyperframe==6.0.1
idna==3.10
imageio==2.36.1
inflate64==1.0.1
iniconfig==2.0.0
invoke==2.2.0
iterative-telemetry==0.0.9
itsdangerous==2.2.0
Jinja2==3.1.5
jmespath==1.0.1
joblib==1.4.2
jsonlines==4.0.0
kaleido==0.2.1
kiwisolver==1.4.8
kombu==5.4.2
lazy_loader==0.4
locust==2.32.6
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
mongoengine==0.29.1
motor==3.6.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multivolumefile==0.2.3
networkx==3.4.2
numpy==1.24.4
oauth2client==4.1.3
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
orjson==3.10.14
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pathtools==0.1.2
pillow==11.1.0
platformdirs==4.3.6
plotly==5.24.1
pluggy==1.5.0
pprintpp==0.4.0
priority==2.0.0
prompt_toolkit==3.0.48
propcache==0.2.1
proto-plus==1.25.0
protobuf==4.25.5
psutil==6.1.1
py-cpuinfo==9.0.0
py7zr==0.22.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pybcj==1.0.3
pycocotools==2.0.8
pycparser==2.22
pycryptodomex==3.21.0
pydantic==2.10.5
pydantic_core==2.27.2
pydash==8.0.4
pydot==3.0.4
PyDrive2==1.21.3
pygit2==1.17.0
Pygments==2.19.1
pygtrie==2.5.0
pymongo==4.9.2
pyOpenSSL==24.2.1
pyparsing==3.2.1
pyppmd==1.1.1
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0
pyzmq==26.2.0
pyzstd==0.16.2
rarfile==4.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
retrying==1.3.4
rich==13.9.4
rsa==4.9
Rtree==1.3.0
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
s3transfer==0.10.4
scikit-image==0.25.0
scikit-learn==1.6.0
scipy==1.15.0
scmrepo==3.3.9
seaborn==0.13.2
semver==3.0.2
sentry-sdk==2.19.2
setproctitle==1.3.4
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.7.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqltrie==0.11.1
sse-starlette==0.10.3
sseclient-py==1.8.0
starlette==0.40.0
strawberry-graphql==0.256.1
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
tensorboard==2.12.0
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
texttable==1.7.0
threadpoolctl==3.5.0
tifffile==2024.12.12
tomlkit==0.13.2
torch==2.1.0
torchmetrics==0.10.0
torchvision==0.16.0
tqdm==4.67.1
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
ultralytics==8.3.59
ultralytics-thop==2.0.13
universal-analytics-python3==1.1.1
uritemplate==4.1.1
urllib3==2.3.0
utils==1.0.2
uvicorn==0.34.0
vine==5.1.0
voluptuous==0.15.2
voxel51-eta==0.13.1
wandb==0.15.0
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.17.0
wsproto==1.2.0
xmltodict==0.14.2
yarl==1.18.3
zc.lockfile==3.0.post1
zope.event==5.0
zope.interface==7.2
